\name{getElem.NewsSource}
\alias{getElem.NewsSource}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Element wrapper function
}
\description{
I still have no idea what this does
}
\usage{
getElem.NewsSource(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Wish I knew
}
}
\details{
Very mysterious.
}
\value{
returns something.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Simon Kiss
}
\note{
Nope
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
list(content = x$content[[x$position]])
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
